Rename ls command to list.
