+++ __tests__/fixtures/list/lockfile-outdated/package.json
