+++ __tests__/fixtures/list/no-args/package.json
