+++ __tests__/fixtures/list/one-arg/package.json
