+++ __tests__/fixtures/request-cache/GET/localhost/.bin
@@ -1,5 +1,5 @@
 HTTP/1.1 200 OK
-Date: Wed, 16 Nov 2016 15:06:25 GMT
+Date: Fri, 18 Nov 2016 20:23:04 GMT
 Connection: close
 Content-Length: 2
