+++ src/cli/aliases.js
@@ -5,7 +5,7 @@ export default ({
   'run-script': 'run',
   c: 'config',
   i: 'install',
-  list: 'ls',
+  ls: 'list',
   ln: 'link',
   rb: 'rebuild',
   runScript: 'run',
@@ -22,8 +22,8 @@ export default ({
   adduser: 'login',
   author: 'owner',
   isntall: 'install',
-  la: 'ls',
-  ll: 'ls',
+  la: 'list',
+  ll: 'list',
   r: 'remove',
   rm: 'remove',
   show: 'info',
